// src/app/models/RepoInfActiEjec.ts
export class RepoInfActiEjec {
    item: number ;
	pendiente: number; 
    n_idOfic: number;
	v_descOfic: string; 
	a_v_usucre: string; 
	v_desacti: string; 
	v_idsubacti: string;
	UM: string; 
	prog_Mes: number; 
	porc_Avance: number;
	prog_Total: number; 
    ava_Total: number;
    ejecMes: number;
	ejecTotal: number;
	v_perfil: string;
  }
  