// src/app/models/request/reporte-request.ts
export class ReportesRequest {
  n_idofic: number;
  n_idcontrati: number;
  v_idacti: string;
  v_subacti: string;
  v_perfil: string;
  v_periodo: string;
  v_usuario: string;
  lstUsuarios: String;
  V_N_VAL_PROG_TOTAL: number;
  V_N_VAL_PROG_MENSUAL: number;
  N_CANT_PERIODOS: number;
  v_ciclo: string;
  v_periodo_inicio: string;
	v_periodo_final: string;
  v_codigo_lector: string;
  v_tipo_inspe: string;
  v_n_item: number;
  d_femision: Date;
  v_n_idempr: number;
}